+++ config/common/aliases.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 return [
     '@root' => dirname(__DIR__, 2),
+    '@src' => '@root/src',
     '@assets' => '@root/public/assets',
     '@assetsUrl' => '@baseUrl/assets',
     '@baseUrl' => '/',
@@ -13,6 +14,4 @@ return [
     '@resources' => '@root/resources',
     '@runtime' => '@root/runtime',
     '@vendor' => '@root/vendor',
-    '@layout' => '@resources/views/layout',
-    '@views' => '@resources/views',
 ];
