+++ config/common/routes.php
@@ -2,13 +2,14 @@
 
 declare(strict_types=1);
 
-use App\Controller\SiteController;
 use Yiisoft\Router\Group;
 use Yiisoft\Router\Route;
 
 return [
     Group::create()
         ->routes(
-            Route::get('/')->action([SiteController::class, 'index'])->name('home'),
+            Route::get('/')
+                ->action(\App\Controller\HomePage\Action::class)
+                ->name('home'),
         ),
 ];
