+++ config/web/di/application.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use App\Handler\NotFoundHandler;
+use App\Handler\NotFound\NotFoundHandler;
 use Yiisoft\Definitions\DynamicReference;
 use Yiisoft\Definitions\Reference;
 use Yiisoft\Injector\Injector;
