+++ src/Controller/HomePage/Action.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Controller\HomePage;
+
+use Psr\Http\Message\ResponseInterface;
+use Yiisoft\Yii\View\Renderer\ViewRenderer;
+
+final readonly class Action
+{
+    public function __construct(
+        private ViewRenderer $viewRenderer,
+    ) {}
+
+    public function __invoke(): ResponseInterface
+    {
+        return $this->viewRenderer->render(__DIR__ . '/template');
+    }
+}
