--- src/Controller/SiteController.php
@@ -1,21 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller;
-
-use Psr\Http\Message\ResponseInterface;
-use Yiisoft\Yii\View\Renderer\ViewRenderer;
-
-final class SiteController
-{
-    public function __construct(private ViewRenderer $viewRenderer)
-    {
-        $this->viewRenderer = $viewRenderer->withControllerName('site');
-    }
-
-    public function index(): ResponseInterface
-    {
-        return $this->viewRenderer->render('index');
-    }
-}
